+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
-import Login from './components/Login.jsx'
-import Home from './components/Home.jsx'
-import Notfound from './components/404.jsx'
+import Login from './views/Login.jsx'
+import Home from './views/Home.jsx'
+import Notfound from './views/404.jsx'
 import Layout from './components/Layout.jsx'
 import Register from './components/Register.jsx'
